Created an util for URL Params extraction and form value exports need to implement this for other schemes
